+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 <!-- loosely based on https://keepachangelog.com/en/1.0.0/ -->
 
-## 0.4.7 - Unreleased
+## 0.4.7 - 2025-10-05
 
 ### Added